moves notify to notifications namespace
