+++ composer.json
@@ -99,6 +99,7 @@
     "autoload": {
         "files": [
             "src/Illuminate/Foundation/helpers.php",
+            "src/Illuminate/Notifications/helpers.php",
             "src/Illuminate/Support/helpers.php"
         ],
         "psr-4": {
